+++ test/generate_rda.R
@@ -1,4 +1,4 @@
-# R script to generate test .rda files
+# R script to generate test .rda and .rds files
 
 df <- data.frame(num = c(1.1, 2.2))
 save(df, file = "data/minimal.rda")
@@ -12,6 +12,9 @@ df["cplx"] <- complex( real = c(1.1, 0.0), imaginary = c(0.5, 1.0) )
 #utf<-c("Ж", "∰")) R handles it, RData doesn"t.
 save(df, file = "data/types.rda")
 save(df, file = "data/types_ascii.rda", ascii = TRUE)
+saveRDS(df, file = "data/types.rds")
+saveRDS(df, file = "data/types_ascii.rds", ascii = TRUE)
+saveRDS(df, file = "data/types_decomp.rds", compress = FALSE)
 
 df[2, ] <- NA
 df[3, ] <- df[2, ]